refactor: merge constructor/method file into one
