--- pkg/broker/configurate_test.go
@@ -1,28 +1,0 @@
-package broker_test
-
-import (
-	"context"
-	"testing"
-	"time"
-)
-
-
-
-
-func TestCreateDeleteTopic(t *testing.T) {
-
-	SetupConfigurator()
-
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 3*time.Second)
-	defer cancelFunc()
-
-	if err := conf.CreateTopic(ctx, topic); err != nil {
-		t.Errorf("CreateTopic() = %v", err)
-	}
-	
-	if err := conf.DeleteTopic(ctx, topic); err != nil {
-		t.Errorf("DeleteTopic() = %v", err)
-	}
-
-	TeardownConfigurator()
-}
